Move Zipkin tracer plugin.
