+++ linkerd/tracer/zipkin/src/main/resources/META-INF/services/io.buoyant.linkerd.TracerInitializer
@@ -1,1 +1,1 @@
-io.l5d.ZipkinTracerInitializer
+io.buoyant.linkerd.tracer.ZipkinTracerInitializer
