+++ linkerd/tracer/zipkin/src/main/scala/io/buoyant/linkerd/tracer/ZipkinTracerInitializer.scala
@@ -1,4 +1,4 @@
-package io.l5d
+package io.buoyant.linkerd.tracer
 
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.twitter.finagle.tracing.{Record, TraceId, Tracer}
@@ -6,12 +6,13 @@ import com.twitter.finagle.zipkin.thrift.{Sampler, ZipkinTracer}
 import io.buoyant.linkerd.{TracerConfig, TracerInitializer}
 
 class ZipkinTracerInitializer extends TracerInitializer {
-  val configClass = classOf[zipkin]
+  val configClass = classOf[ZipkinConfig]
+  override def configId = "io.l5d.zipkin"
 }
 
 object ZipkinTracerInitializer extends ZipkinTracerInitializer
 
-case class zipkin(
+case class ZipkinConfig(
   host: Option[String],
   port: Option[Int],
   sampleRate: Option[Double]
@@ -30,3 +31,4 @@ case class zipkin(
     def record(r: Record) = underlying.record(r)
   }
 }
+
